release v1.5.2
- add cnpapi and host collectors

+++ build_image.sh
@@ -16,14 +16,16 @@
 curpath=$(dirname "$0")
 cd "$curpath" || exit 1
 
-: "${TAG:=v1.5.1}"
+: "${TAG:=v1.5.2}"
 : "${ARCH:=amd64}"
 : "${LIBCNDEV:=/usr/local/neuware/lib64/libcndev.so}"
+: "${LIBCNPAPI:=/usr/local/neuware/lib64/libcnpapi.so}"
 
 echo "Build environ (Can be overridden):"
 echo "TAG       = $TAG"
 echo "ARCH      = $ARCH"
 echo "LIBCNDEV  = $LIBCNDEV"
+echo "LIBCNPAPI  = $LIBCNPAPI"
 echo "APT_PROXY = $APT_PROXY"
 echo "GOPROXY   = $GOPROXY"
 
@@ -48,6 +50,12 @@ if [[ ! -f "$LIBCNDEV" ]]; then
 	exit 1
 fi
 
+if [[ ! -f "$LIBCNPAPI" ]]; then
+	echo "Can't find libcnpapi.so at $LIBCNPAPI."
+	echo "If you want to scrape cnpapi metrics, please install Cambricon neuware, or set LIBCNPAPI environ to path of libcnpapi.so"
+	echo "Else, ignore this message."
+fi
+
 case $ARCH in
 amd64)
 	file_arch=x86-64
@@ -65,7 +73,13 @@ if ! file "$LIBCNDEV" --dereference | grep -q "$file_arch"; then
 	exit 1
 fi
 
+if [[ -f "$LIBCNPAPI" ]] && ! file "$LIBCNPAPI" --dereference | grep -q "$file_arch"; then
+	echo "$LIBCNPAPI is not for $ARCH"
+	exit 1
+fi
+
 cp "$LIBCNDEV" "$curpath/libs/linux/$ARCH/libcndev.so"
+[[ -f "$LIBCNPAPI" ]] && cp "$LIBCNPAPI" "$curpath/libs/linux/$ARCH/libcnpapi.so"
 
 echo "Building Cambricon MLU Exporter docker image."
 
@@ -91,3 +105,4 @@ fi
 
 echo "Image is saved at ./image/cambricon-mlu-exporter-$ARCH.tar"
 rm -f "$curpath/libs/linux/$ARCH/libcndev.so"
+rm -f "$curpath/libs/linux/$ARCH/libcnpapi.so"
